Move some slow tests from `test_targeting.py` into `nocover/`
