+++ hypothesis-python/tests/nocover/test_targeting.py
@@ -15,7 +15,7 @@
 
 import pytest
 
-from hypothesis import Phase, given, settings, strategies as st, target
+from hypothesis import Phase, given, seed, settings, strategies as st, target
 
 pytest_plugins = "pytester"
 
@@ -61,3 +61,36 @@ def test_target_returns_value(a, b):
     difference = target(abs(a - b))
     assert difference == abs(a - b)
     assert isinstance(difference, int)
+
+
+def test_targeting_can_be_disabled():
+    strat = st.lists(st.integers(0, 255))
+
+    def score(enabled):
+        result = [0]
+        phases = [Phase.generate]
+        if enabled:
+            phases.append(Phase.target)
+
+        @seed(0)
+        @settings(database=None, max_examples=200, phases=phases)
+        @given(strat)
+        def test(ls):
+            score = float(sum(ls))
+            result[0] = max(result[0], score)
+            target(score)
+
+        test()
+        return result[0]
+
+    assert score(enabled=True) > score(enabled=False)
+
+
+def test_issue_2395_regression():
+    @given(d=st.floats().filter(lambda x: abs(x) < 1000))
+    @settings(max_examples=1000, database=None)
+    @seed(93962505385993024185959759429298090872)
+    def test_targeting_square_loss(d):
+        target(-((d - 42.5) ** 2.0))
+
+    test_targeting_square_loss()
